// mobile/src/api/generated/schemas/workoutExerciseResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * GrowIt.API | v1
 * OpenAPI spec version: 1.0.0
 */
import type { SetResponse } from './setResponse';

export interface WorkoutExerciseResponse {
  id: string;
  /** @nullable */
  exerciseId: string | null;
  exerciseName: string;
  /** @nullable */
  category: string | null;
  /** @nullable */
  targetSets: number | null;
  /** @nullable */
  targetReps: number | null;
  /** @nullable */
  restSeconds: number | null;
  orderIndex: number;
  sets: SetResponse[];
}